+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Keyboard shortcut to refresh the controller (CTRL-R)
 - Button to refresh the controller (next to the Interface selection dropdown)
 - Entity Identify notifications (from entity to controller)
+- Entity and Full Network export as readable json
+- Strings descriptor displayed
 
 ### Changed
 - Only displaying _Ethernet_ kind interfaces
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Possible deadlock when trying to match stream formats
 - [Random crash during application exit](https://github.com/christophe-calmejane/Hive/issues/56)
+- StreamInput counters not properly displayed (as sub-nodes of the Counter node)
 
 ## [1.0.11] - 2019-02-14
 ### Added